// src/shared/constants/enums.ts
export enum UserRole {
  ADMIN = 'Admin',
  MANAGER = 'Manager',
  STAFF = 'Staff',
}

export enum MovementType {
  INBOUND = 'INBOUND',
  OUTBOUND = 'OUTBOUND',
  TRANSFER_OUT = 'TRANSFER_OUT',
  TRANSFER_IN = 'TRANSFER_IN',
  ADJUSTMENT = 'ADJUSTMENT',
}

export enum ReferenceType {
  PURCHASE = 'PURCHASE',
  SALE = 'SALE',
  COUNT = 'COUNT',
  TRANSFER = 'TRANSFER',
}